Change profTitle to not be abbreviated.
I just propose to remove the period after of the professor's title.
With the period there, people are forced to have it look weird or
abbreviate it, while if it isn't there, they can choose to abbreviate or
not. (I also think the format doesn't want it abbreviated but it doesn't
say it directly as far as I can tell)