Remove `profTitle`
removing the default title and just having the user write it
makes more sense for this. The issue with the default title, is
that it is slightly simplier for when you use the default title,
but it makes it more annoying anytime you don't use it.

In the case where you want to remove the title, there will be a leading
space infront of their name, and that's also annoying.